services/common/auth/requirements.txt
# CoLink Common Auth Module Dependencies

fastapi>=0.104.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
httpx>=0.25.0
python-multipart>=0.0.6
